Implement method execution duration measurement. Add use-stopwatch command line parameter.

+++ FileCabinetApp/CommandHandlers/PurgeCommandHandler.cs
@@ -9,13 +9,16 @@ namespace FileCabinetApp.CommandHandlers
     /// </summary>
     public class PurgeCommandHandler : ServiceCommandHandlerBase
     {
+        private bool useStopWatch;
+
         /// <summary>
         /// Initializes a new instance of the <see cref="PurgeCommandHandler"/> class.
         /// </summary>
         /// <param name="service">The IFileCabinetService service.</param>
-        public PurgeCommandHandler(IFileCabinetService service)
+        public PurgeCommandHandler(IFileCabinetService service, bool useStopWatch)
             : base(service)
         {
+            this.useStopWatch = useStopWatch;
         }
 
         /// <summary>
@@ -29,7 +32,17 @@ namespace FileCabinetApp.CommandHandlers
                 if (this.service is FileCabinetFilesystemService)
                 {
                     var totalAmountOfRecords = this.service.GetStat().total;
-                    this.service.Purge();
+
+                    if (this.useStopWatch)
+                    {
+                        ServiceMeter serviceMeter = new ServiceMeter(this.service);
+                        serviceMeter.Purge();
+                    }
+                    else
+                    {
+                        this.service.Purge();
+                    }
+
                     var purgedRecords = totalAmountOfRecords - this.service.GetStat().total;
                     Console.WriteLine($"Data file processing is completed: {purgedRecords} of {totalAmountOfRecords} were purged.");
                 }
